+++ src/test/java/org/recap/ncip/CancelRequestItemUT.java
@@ -1,4 +1,4 @@
-package org.recap;
+package org.recap.ncip;
 
 import org.extensiblecatalog.ncip.v2.service.CancelRequestItemResponseData;
 import org.extensiblecatalog.ncip.v2.service.ItemId;
@@ -7,7 +7,7 @@ import org.extensiblecatalog.ncip.v2.service.ProblemType;
 import org.extensiblecatalog.ncip.v2.service.RequestId;
 import org.junit.Test;
 import org.mockito.InjectMocks;
-import org.recap.ncip.CancelRequestItem;
+import org.recap.BaseTestCaseUT;
 
 import java.util.Arrays;
 import java.util.Collections;
