+++ src/test/java/org/recap/ncip/CheckInItemUT.java
@@ -1,4 +1,4 @@
-package org.recap;
+package org.recap.ncip;
 
 import org.extensiblecatalog.ncip.v2.service.CheckInItemResponseData;
 import org.extensiblecatalog.ncip.v2.service.ItemId;
@@ -8,7 +8,7 @@ import org.extensiblecatalog.ncip.v2.service.ProblemType;
 import org.json.JSONObject;
 import org.junit.Test;
 import org.mockito.InjectMocks;
-import org.recap.ncip.CheckinItem;
+import org.recap.BaseTestCaseUT;
 
 import java.util.Arrays;
 import java.util.Collections;
