+++ src/test/java/org/recap/ncip/CheckoutItemUT.java
@@ -1,4 +1,4 @@
-package org.recap;
+package org.recap.ncip;
 
 import org.extensiblecatalog.ncip.v2.service.CheckOutItemResponseData;
 import org.extensiblecatalog.ncip.v2.service.ItemId;
@@ -8,7 +8,7 @@ import org.extensiblecatalog.ncip.v2.service.UserId;
 import org.json.JSONObject;
 import org.junit.Test;
 import org.mockito.InjectMocks;
-import org.recap.ncip.CheckoutItem;
+import org.recap.BaseTestCaseUT;
 
 import java.util.Arrays;
 import java.util.Collections;
